add GameStartForm component
