+++ components/GameStartForm/styles.module.css
@@ -1,16 +1,4 @@
-.homepageTitle {
-  font-size: 80px;
-  color: var(--yellow);
-  text-shadow: 0 1px 0 #3c391e, 0 2px 0 #3c391e, 0 3px 0 #3c391e,
-    0 4px 0 #3c391e, 0 5px 0 #3c391e, 0 20px 30px rgba(0, 0, 0, 0.5);
-}
-
-.difficultyHeading {
-  font-size: 40px;
-  text-shadow: 2px 2px 4px black;
-}
-
-/* .homepageForm {
+.homepageForm {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -64,4 +52,4 @@
   padding: 12px 32px;
   margin: 20px 0;
   text-shadow: 1px 1px 2px black;
-} */
+}
